fp16

+++ llava/eval/maya/eval_utils.py
@@ -48,7 +48,7 @@ def load_maya_model(model_base: str, model_path : str, projector_path : Optional
 
     device_map = 'auto'
     kwargs = {"device_map": device_map}
-    kwargs['torch_dtype'] = torch.float32
+    kwargs['torch_dtype'] = torch.float16
     #kwargs['attn_implementation'] = 'flash_attention_2'
 
     ## Instantiating tokenizer and model base
